Relaxing git ignore rules, license and Gradle upgrade
Several small changes are included here.
First, this will change the git ignore rules to `intellij` instead of
`intellij+all`. This will help with configurations, e.g. license blocks.

Next in line is the license. This project is under AGPL and the `COPYING`
file reflects that.

Last but not least, Gradle is updated to version 4.10.